Moved plugin activation class into own folder-structure
